+++ app/components/LoadingIndicatorDots/LoadingIndicatorDots.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { View } from 'react-native'
-import AnimatedEllipsis from 'rn-animated-ellipsis'
+import AnimatedEllipsis from '../AnimatedEllipsis/AnimatedEllipsis'
 import { useDynamicStyles } from '../../hooks/useDynamicStyles'
 
 import dynamicStyleSheet from './LoadingIndicatorDots.styles'
@@ -14,7 +14,6 @@ export default function LoadingIndicatorDots(): React.ReactElement {
         style={styles.loadingDots}
         minOpacity={0.4}
         animationDelay={200}
-        useNativeDriver={true}
       />
     </View>
   )